修改错误码信息、修改锁仓epoch值太小问题

+++ browser-common/src/main/java/com/platon/browser/param/RestrictingCreateParam.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.experimental.Accessors;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -34,7 +35,7 @@ public class RestrictingCreateParam extends TxParam{
         /**
          * 表示结算周期的倍数。与每个结算周期出块数的乘积表示在目标区块高度上释放锁定的资金。Epoch * 每周期的区块数至少要大于最高不可逆区块高度
          */
-        private Integer epoch;
+        private BigInteger epoch;
 
         /**
          * 表示目标区块上待释放的金额
