+++ browser-service/src/main/java/com/platon/browser/res/transaction/TransactionDetailsRPPlanResp.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.platon.browser.config.CustomLatSerializer;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  *  交易详情锁仓子结构体返回对象
@@ -13,13 +14,13 @@ import java.math.BigDecimal;
  *  @data 2019年8月31日
  */
 public class TransactionDetailsRPPlanResp {
-	private Integer epoch;         //锁仓周期
+	private BigInteger epoch;         //锁仓周期
     private BigDecimal amount;      //锁定金额
     private Long blockNumber;   //锁仓周期对应快高  结束周期 * epoch  
-	public Integer getEpoch() {
+	public BigInteger getEpoch() {
 		return epoch;
 	}
-	public void setEpoch(Integer epoch) {
+	public void setEpoch(BigInteger epoch) {
 		this.epoch = epoch;
 	}
 	@JsonSerialize(using = CustomLatSerializer.class)
